+++ packages/fluentui/docs/src/examples/components/Attachment/commonScreenerSteps.ts
@@ -1,8 +1,8 @@
-import { Attachment } from '@fluentui/react-northstar';
+import { Attachment, AttachmentAction } from '@fluentui/react-northstar';
 
 const selectors = {
   root: `.${Attachment.className}`,
-  action: `.${Attachment.slotClassNames.action}`,
+  action: `.${AttachmentAction.className}`,
 };
 
 const getScreenerSteps = (): ScreenerSteps => [