+++ packages/fluentui/react-bindings/src/telemetry/useTelemetry.ts
@@ -1,5 +1,5 @@
-import useComposeOptions from '../compose/useComposeOptions';
 import { Telemetry, UseTelemetryResult } from './types';
+import useComposeOptions from '../compose/useComposeOptions';
 
 export const getTelemetry = (displayName: string, telemetry: Telemetry | undefined): UseTelemetryResult => {
   let start: number = -1;
@@ -34,8 +34,9 @@ export const getTelemetry = (displayName: string, telemetry: Telemetry | undefin
 
 const useTelemetry = (displayName: string, telemetry: Telemetry | undefined): UseTelemetryResult => {
   const composeOptions = useComposeOptions();
+  const telemetryName = composeOptions?.displayNames[composeOptions?.displayNames.length - 1] || displayName;
 
-  return getTelemetry(composeOptions?.displayNames[0] || displayName, telemetry);
+  return getTelemetry(telemetryName, telemetry);
 };
 
 export default useTelemetry;