+++ packages/fluentui/react-northstar/src/themes/teams/components/Attachment/attachmentStyles.ts
@@ -1,12 +1,11 @@
 import { ComponentSlotStylesPrepared, ICSSInJSStyle } from '@fluentui/styles';
 
 import { AttachmentStylesProps } from '../../../../components/Attachment/Attachment';
+import AttachmentAction from '../../../../components/Attachment/AttachmentAction';
 import { AttachmentVariables } from './attachmentVariables';
 import { pxToRem } from '../../../../utils';
 import SvgIcon from '../../../../components/SvgIcon/SvgIcon';
 import getBorderFocusStyles from '../../getBorderFocusStyles';
-import getIconFillOrOutlineStyles from '../../getIconFillOrOutlineStyles';
-import Button from '../../../../components/Button/Button';
 
 const attachmentStyles: ComponentSlotStylesPrepared<AttachmentStylesProps, AttachmentVariables> = {
   root: ({ props: p, variables: v, theme: { siteVariables } }): ICSSInJSStyle => {
@@ -33,6 +32,17 @@ const attachmentStyles: ComponentSlotStylesPrepared<AttachmentStylesProps, Attac
 
       ...borderFocusStyles,
 
+      '& .ui-attachment__progress': {
+        transition: 'width 0.2s',
+        position: 'absolute',
+        display: 'block',
+        bottom: 0,
+        left: 0,
+        maxWidth: '100%',
+        height: pxToRem(v.progressHeight),
+        background: v.progressColor,
+      },
+
       ...(p.actionable && {
         cursor: 'pointer',
 
@@ -41,7 +51,7 @@ const attachmentStyles: ComponentSlotStylesPrepared<AttachmentStylesProps, Attac
           backgroundColor: v.focusBackgroundColor,
           color: v.focusColor,
 
-          [`& .${Button.className}`]: {
+          [`& .${AttachmentAction.className}`]: {
             color: v.siblingsFocusColor,
           },
 
@@ -54,7 +64,7 @@ const attachmentStyles: ComponentSlotStylesPrepared<AttachmentStylesProps, Attac
           background: v.backgroundColorHover,
           color: v.textColorHover,
 
-          [`& .${Button.className}`]: {
+          [`& .${AttachmentAction.className}`]: {
             color: v.siblingsHoverColor,
           },
 
@@ -65,76 +75,6 @@ const attachmentStyles: ComponentSlotStylesPrepared<AttachmentStylesProps, Attac
       }),
     };
   },
-
-  content: (): ICSSInJSStyle => ({
-    flex: 1,
-  }),
-
-  header: ({ variables: v }): ICSSInJSStyle => ({
-    fontSize: v.headerFontSize,
-    fontWeight: v.headerFontWeight,
-    lineHeight: v.headerLineHeight,
-  }),
-
-  description: ({ variables: v }): ICSSInJSStyle => ({
-    display: 'block',
-    fontSize: v.descriptionFontSize,
-    fontWeight: v.descriptionFontWeight,
-    lineHeight: v.descriptionLineHeight,
-  }),
-
-  icon: ({ variables: v }): ICSSInJSStyle => ({
-    flex: '0 0 auto',
-    height: v.iconSize,
-    width: v.iconSize,
-    [`& .${SvgIcon.className}`]: {
-      height: '100%',
-      width: '100%',
-      '& svg': {
-        height: '100%',
-        width: '100%',
-      },
-    },
-    marginRight: v.iconSpace,
-  }),
-
-  action: ({ props: p, variables: v, theme: { siteVariables } }): ICSSInJSStyle => {
-    const iconFilledStyles = getIconFillOrOutlineStyles({ outline: false });
-    const borderFocusStyles = getBorderFocusStyles({
-      variables: siteVariables,
-      borderRadius: v.borderRadius,
-    });
-
-    return {
-      [`& .${SvgIcon.className}`]: {
-        color: v.textColor, // this breaks the color change on hover
-      },
-
-      ...getIconFillOrOutlineStyles({ outline: true }),
-
-      ':hover': {
-        ...iconFilledStyles,
-        background: 'transparent',
-      },
-
-      ':focus': borderFocusStyles[':focus'],
-      ':focus-visible': {
-        ...iconFilledStyles,
-        ...borderFocusStyles[':focus-visible'],
-      },
-    };
-  },
-
-  progress: ({ props: p, variables: v }): ICSSInJSStyle => ({
-    transition: 'width 0.2s',
-    position: 'absolute',
-    display: 'block',
-    bottom: 0,
-    left: 0,
-    maxWidth: '100%',
-    height: pxToRem(v.progressHeight),
-    background: v.progressColor,
-  }),
 };
 
 export default attachmentStyles;